Rename integrator app to integrate
