+++ integrate/test_odoo.py
@@ -45,7 +45,7 @@ class OdooTests(TestCase):
 
         with connection.cursor() as cursor:
             cursor.execute(
-                "SELECT password FROM integrator_odooinstance WHERE id = %s",
+                "SELECT password FROM integrate_odooinstance WHERE id = %s",
                 [self.instance.pk],
             )
             raw = cursor.fetchone()[0]
@@ -55,7 +55,7 @@ class OdooTests(TestCase):
     def test_connection_success(self, mock_proxy):
         mock_srv = mock_proxy.return_value
         mock_srv.authenticate.return_value = 1
-        response = self.client.post(reverse("integrator:odoo-test", args=[self.instance.pk]))
+        response = self.client.post(reverse("integrate:odoo-test", args=[self.instance.pk]))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json()["detail"], "success")
 
@@ -63,7 +63,7 @@ class OdooTests(TestCase):
     def test_connection_invalid(self, mock_proxy):
         mock_srv = mock_proxy.return_value
         mock_srv.authenticate.return_value = False
-        response = self.client.post(reverse("integrator:odoo-test", args=[self.instance.pk]))
+        response = self.client.post(reverse("integrate:odoo-test", args=[self.instance.pk]))
         self.assertEqual(response.status_code, 401)
 
 
@@ -96,7 +96,7 @@ class OdooAdminTests(TestCase):
     def test_admin_action(self, mock_proxy):
         mock_srv = mock_proxy.return_value
         mock_srv.authenticate.return_value = 1
-        url = reverse("admin:integrator_odooinstance_changelist")
+        url = reverse("admin:integrate_odooinstance_changelist")
         resp = self.client.post(
             url,
             {
