+++ release/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from integrator.models import Entity
+from integrate.models import Entity
 from . import Package, Credentials, DEFAULT_PACKAGE
